Split the GNOME extension into compositor-only ES modules
Breaks the extension.js monolith into one module per primitive, keeping
the entrypoint thin: grid.js (overlay + input injection), windows.js
(window/workspace ops), screenshot.js (capture), and indicator.js (tray
+ Quick Settings). extension.js now owns only the Extension lifecycle and
the D-Bus interface and handler wiring, importing the modules.

Adds the new files to EXTENSION_ASSETS and the setuptools package-data so
they ship, and switches the eslint config and lint-js recipe to a
gnome@easyspeak.dev/*.js glob so future modules are covered without
maintaining parallel lists.

Part of #75.
